World Class for world Generation
Created a map generator that will take a list of values, in a range between 0 and 5, and will output a tile on a new list for each value of the list. This will then be updated on the screen. This provides the platforms and ground for the game. The next update will be collisions, to allow the platforming to work, as well as changing the sprite for the player to make it visually easier to determine jumps.